removed logic regarding users from AdminController to ManageUsersController
